# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/candidates": {
      "get": {
        "operationId": "getCandidates",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by. Defaults to most recently created.",
            "schema": {
              "type": "string",
              "enum": [
                "displayId",
                "fullName",
                "familyName",
                "givenName",
                "yearsExperience",
                "lastContactedAt",
                "status"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search across fullName, email, displayId, mobile, city, country, currentPosition, currentCompany, and industry/role type name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statuses",
            "required": false,
            "in": "query",
            "description": "Filter by status (one or more). Omit for all statuses.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "COLD",
                  "WARM",
                  "HOT",
                  "PLACED"
                ]
              }
            }
          },
          {
            "name": "industryIds",
            "required": false,
            "in": "query",
            "description": "Filter by industry ID(s) (see /industries)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "roleTypeIds",
            "required": false,
            "in": "query",
            "description": "Filter by role type ID(s) (see /candidate-role-types)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "specializationIds",
            "required": false,
            "in": "query",
            "description": "Filter by specialization ID(s) (see /specializations)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "skills",
            "required": false,
            "in": "query",
            "description": "Filter by skill tag(s) (exact match against the free-entry skills list)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "consultantIds",
            "required": false,
            "in": "query",
            "description": "Filter by owning consultant ID(s)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "submissionStatuses",
            "required": false,
            "in": "query",
            "description": "Filter by submission status (has at least one submission with this status)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "SUBMITTED",
                  "INTERVIEWING",
                  "REJECTED",
                  "PLACED"
                ]
              }
            }
          },
          {
            "name": "placementStatuses",
            "required": false,
            "in": "query",
            "description": "Filter by placement status (has at least one placement with this status)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "COMPLETED",
                  "FAILED"
                ]
              }
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "description": "Filter by location — matches city OR country (contains, case-insensitive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currentCompany",
            "required": false,
            "in": "query",
            "description": "Filter by current company (contains, case-insensitive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currentPosition",
            "required": false,
            "in": "query",
            "description": "Filter by current position (contains, case-insensitive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "yearsExperienceMin",
            "required": false,
            "in": "query",
            "description": "Minimum years of experience (inclusive)",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "yearsExperienceMax",
            "required": false,
            "in": "query",
            "description": "Maximum years of experience (inclusive)",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "lastContactedFrom",
            "required": false,
            "in": "query",
            "description": "Only candidates last contacted on/after this date (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastContactedTo",
            "required": false,
            "in": "query",
            "description": "Only candidates last contacted on/before this date (ISO 8601)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated candidates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCandidatesEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List candidates (paginated, filterable, sortable)",
        "tags": [
          "Candidates"
        ]
      },
      "post": {
        "operationId": "createCandidate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Candidate created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new candidate",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/saved-searches": {
      "get": {
        "operationId": "getCandidateSavedSearches",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Saved searches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateSavedSearchEntity"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the caller's saved candidate searches",
        "tags": [
          "Candidates"
        ]
      },
      "post": {
        "operationId": "createCandidateSavedSearch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateSavedSearchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Saved search created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateSavedSearchEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Save the current candidate search/filter state",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/saved-searches/{id}": {
      "delete": {
        "operationId": "deleteCandidateSavedSearch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Saved search deleted"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete one of the caller's saved searches",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/by-display-id/{displayId}": {
      "get": {
        "operationId": "getCandidateByDisplayId",
        "parameters": [
          {
            "name": "displayId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get candidate by display ID (CDD-XXXX)",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}": {
      "get": {
        "operationId": "getCandidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get candidate by ID",
        "tags": [
          "Candidates"
        ]
      },
      "patch": {
        "operationId": "updateCandidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCandidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Candidate updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a candidate",
        "tags": [
          "Candidates"
        ]
      },
      "delete": {
        "operationId": "deleteCandidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Candidate soft-deleted"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a candidate (recoverable)",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}/pipeline-timeline": {
      "get": {
        "operationId": "getCandidatePipelineTimeline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pipeline events, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PipelineTimelineEventEntity"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "This candidate's submission/stage history across every job order",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}/notes": {
      "post": {
        "operationId": "addCandidateNote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCandidateNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Candidate updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Append a note to a candidate's timeline",
        "tags": [
          "Candidates"
        ]
      }
    },
...
  },
  "info": {
    "title": "Linktal Recruitment API",
    "description": "API for Linktal Recruitment OS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginatedCandidatesEntity": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateEntity"
            }
          },
          "total": {
            "type": "number",
            "description": "Total rows matching the filters",
            "example": 342
          },
          "page": {
            "type": "number",
            "description": "Current page (1-based)",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "description": "Items per page",
            "example": 20
          },
          "pageCount": {
            "type": "number",
            "description": "Total number of pages",
            "example": 18
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "pageSize",
          "pageCount"
        ]
      },
...
      "AddCandidateNoteDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Note content",
            "example": "Called to confirm availability for the panel interview."
          }
        },
        "required": [
          "content"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND",
            "description": "Stable machine-readable code"
          },
          "message": {
            "type": "string",
            "example": "Candidate abc not found"
          },
          "details": {
            "nullable": true,
            "description": "Field-level messages (e.g. validation errors)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string",
            "example": "/api/candidates/abc"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-09T10:20:30.000Z"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "details",
          "path",
          "timestamp"
        ]
      }
    }
  }
}